+++ src/core/utils/index.ts
@@ -1,6 +1,9 @@
 export * from './Queue';
 export * from './Result';
 export * from './logger';
+export * from './makeEventStore';
+export * from './makeEventStoreRepo';
+export * from './makeEventStoreTransactionalRepo';
 export * from './stableStringify';
 export * from './validators';
 export * from './withDelay';
