--- src/infra/eventStore/index.ts
@@ -1,3 +1,0 @@
-export * from './BaseEventStore';
-export * from './makeEventStoreRepo';
-export * from './makeEventStoreTransactionalRepo';
