+++ src/infra/sequelize/repos/projectClaimRepo.ts
@@ -5,8 +5,8 @@ import {
   TransactionalRepository,
   UniqueEntityID,
 } from '../../../core/domain'
+import { makeEventStoreRepo, makeEventStoreTransactionalRepo } from '../../../core/utils'
 import { makeProjectClaim, ProjectClaim } from '../../../modules/projectClaim'
-import { makeEventStoreRepo, makeEventStoreTransactionalRepo } from '../../eventStore'
 
 export const makeProjectClaimRepo = (
   eventStore: EventStore
