+++ src/infra/sequelize/repos/projectRepo.ts
@@ -5,10 +5,10 @@ import {
   TransactionalRepository,
   UniqueEntityID,
 } from '../../../core/domain'
+import { makeEventStoreRepo, makeEventStoreTransactionalRepo } from '../../../core/utils'
 import { appelsOffreStatic } from '../../../dataAccess/inMemory'
 import { AppelOffre } from '../../../entities'
 import { BuildProjectIdentifier, makeProject, Project } from '../../../modules/project'
-import { makeEventStoreRepo, makeEventStoreTransactionalRepo } from '../../eventStore'
 
 const appelsOffres: Record<AppelOffre['id'], AppelOffre> = appelsOffreStatic.reduce(
   (map, appelOffre) => {
